# chat-api/src/main/java/uz/zohidjon/chatapi/message/MessageRequest.java
package uz.zohidjon.chatapi.message;

import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class MessageRequest {

    private String content;
    private String senderId;
    private String receiverId;
    private MessageType type;
    private String chatId;
}