# chatapp-ui/src/app/services/models/chat-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChatResponse {
  id?: string;
  lastMessage?: string;
  lastMessageTime?: string;
  name?: string;
  receiverId?: string;
  recipientOnline?: boolean;
  senderId?: string;
  unreadCount?: number;
}