# chatapp-ui/src/app/services/models/user-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserResponse {
  email?: string;
  firstname?: string;
  id?: string;
  lastSeen?: string;
  lastname?: string;
  online?: boolean;
}